+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This documents the changes of every release starting of 0.1.0.
 * Support additional constant switches for test runs
 * Support `uv run` to run pytest (#47)
 * Support to specify python version, when uv is used (#49)
+* Support to specify a specific version of a package, when uv is used
 * Support preselection of switches for test commands (#42)
 * Support for non persistent flags (#41)